fix(onboard): revert Telegram non-interactive skip
Reverts b13e4f4a5 (#4307).

That change caused fake-token E2E paths to drop Telegram during

non-interactive onboarding, failing provider, registry, channel-list,

and token-rotation assertions.
